Refactor in receptionn

+++ Sources/Core/MedicalOffice.Application/Dtos/CashDTO/CashListDTO.cs
@@ -13,7 +13,7 @@ public class CashListDTO :BaseDto<Guid>
     /// <summary>
     /// پرداختی
     /// </summary>
-    public long Cost { get; set; }
+    public float Cost { get; set; }
     /// <summary>
     /// آیدی پذیرش
     /// </summary>
