+++ Sources/Core/MedicalOffice.Application/Dtos/CashDTO/UpdateCashCheckDTO.cs
@@ -12,7 +12,7 @@ public class UpdateCashCheckDTO : BaseDto<Guid>
     /// <summary>
     /// پرداختی
     /// </summary>
-    public long Cost { get; set; }
+    public float Cost { get; set; }
     /// <summary>
     /// شماره حساب
     /// </summary>
