+++ Sources/Core/MedicalOffice.Application/Dtos/InsuranceDTO/InsuranceListDTO.cs
@@ -14,7 +14,7 @@ public class InsuranceListDTO : BaseDto<Guid>
     /// <summary>
     /// کد بیمه
     /// </summary>
-    public long InsuranceCode { get; set; }
+    public float InsuranceCode { get; set; }
 
     /// <summary>
     /// درصد بیمه
