+++ Sources/Core/MedicalOffice.Application/Dtos/MemberShipServiceDTO/UpdateMemberShipServiceDTO.cs
@@ -23,6 +23,6 @@ namespace MedicalOffice.Application.Dtos.MemberShipServiceDTO
         /// <summary>
         /// تخفیف
         /// </summary>
-        public long Discount { get; set; }
+        public float Discount { get; set; }
     }
 }
