+++ Sources/Core/MedicalOffice.Application/Dtos/Tariff/TariffListDTO.cs
@@ -28,7 +28,7 @@ public class TariffListDTO : BaseDto<Guid>
     /// <summary>
     /// کد بیمه
     /// </summary>
-    public long? TariffInsuranceCode { get; set; }
+    public float? TariffInsuranceCode { get; set; }
 
     /// <summary>
     /// مبلغ تعرفه
